Add files via upload
Completed a Windows-based application using the tkinter module for school.The tkinter module is a Python library for creating graphical user interfaces (GUIs). It is a cross-platform library, meaning that it can be used to create GUIs for Windows, Mac, and Linux. Tkinter is a relatively simple library to use, making it a good choice for beginners.
 In the future, I can develop,
 1. Add more styles to my application. This could include changing the colors, fonts, and layout of the application. I could also add images and other graphical elements to make the application more visually appealing.
2. Make my application more like a website. This could involve adding features such as a navigation bar, a search bar, and a contact form. I could also make the application responsive so that it can be viewed on different devices.
3. Add more functionality to my application.